Refactor namespaces in audit-related configurations and models for better organization
