+++ tests/Audit/Configurations/Cases/AuditEntityTypeConfiguration.cs
@@ -3,9 +3,9 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-using Wangkanai.Domain.Models;
+using Wangkanai.Audit.Models;
 
-namespace Wangkanai.Domain.Configurations;
+namespace Wangkanai.Audit.Configurations.Cases;
 
 public class AuditEntityTypeConfiguration : IEntityTypeConfiguration<AuditEntity>
 {
