+++ tests/Audit/Configurations/Cases/IntEntityTypeConfiguration.cs
@@ -5,7 +5,7 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 using Wangkanai.Audit.Models;
 
-namespace Wangkanai.Domain.Configurations;
+namespace Wangkanai.Audit.Configurations.Cases;
 
 public class IntEntityTypeConfiguration : IEntityTypeConfiguration<IntEntity>
 {
